requirements.txt
kaggle~=1.6.14
python-dotenv~=1.0.1
azure-cognitiveservices-vision-customvision~=3.1.0
pyyaml~=6.0.1
globox~=2.4.5
roboflow~=1.1.33